adjust tests to separate impl details
