+++ test/parallel_api/iterator/permutation_iterator.pass.cpp
@@ -17,6 +17,12 @@
 
 #include "permutation_iterator_common.h"
 
+#if TEST_DPCPP_BACKEND_PRESENT
+#   include "support/utils_sycl.h"
+#endif
+
+using namespace TestUtils;
+
 int
 main()
 {
@@ -36,23 +42,37 @@ main()
     auto permItBegin = dpl::make_permutation_iterator(countingItBegin, kDefaultIndexStepOp);
     auto permItEnd = permItBegin + perm_size_expected;
 
-    //transform_iterator is not trivially_copyable, as it defines a copy assignment operator which does not copy
-    // its unary functor.  permutation_iterator is based on transform_iterator and therefore is also not
-    // trivially_copyable.  This makes permutation_iterator's device_copyable trait deprecated with sycl 2020.
-    EXPECT_TRUE(check_if_device_copyable_by_sycl2020_or_by_old_definition <decltype(permItBegin)>,
-                "permutation_iterator (counting_iterator) is not properly copyable");
+   static_assert(sycl::is_device_copyable_v<decltype(permItBegin)>,
+                "permutation_iterator (counting_iterator) is not device copyable");
+
+    static_assert(
+        sycl::is_device_copyable_v<
+            oneapi::dpl::permutation_iterator<constant_iterator_device_copyable, constant_iterator_device_copyable>>,
+        "permutation_iterator is not device copyable with device copyable types");
+    static_assert(sycl::is_device_copyable_v<
+                      oneapi::dpl::permutation_iterator<constant_iterator_device_copyable, noop_device_copyable>>,
+                  "permutation_iterator is not device copyable with device copyable types");
+
+    static_assert(!sycl::is_device_copyable_v<
+                      oneapi::dpl::permutation_iterator<constant_iterator_non_device_copyable, noop_device_copyable>>,
+                  "permutation_iterator is device copyable with non device copyable types");
+    static_assert(!sycl::is_device_copyable_v<oneapi::dpl::permutation_iterator<int*, noop_non_device_copyable>>,
+                  "permutation_iterator is device copyable with non device copyable types");
+    static_assert(
+        !sycl::is_device_copyable_v<oneapi::dpl::permutation_iterator<int*, constant_iterator_non_device_copyable>>,
+        "permutation_iterator is device copyable with non device copyable types");
 
     const std::size_t perm_size_result = std::distance(permItBegin, permItEnd);
     EXPECT_EQ(perm_size_expected, perm_size_result,
               "Wrong result of std::distance<permutationIterator1, permutationIterator2)");
 
     std::vector<int> resultCopy(perm_size_result);
-    auto itCopiedDataEnd = dpl::copy(TestUtils::default_dpcpp_policy, permItBegin, permItEnd, resultCopy.begin());
+    auto itCopiedDataEnd = dpl::copy(default_dpcpp_policy, permItBegin, permItEnd, resultCopy.begin());
     EXPECT_EQ(true, resultCopy.end() == itCopiedDataEnd, "Wrong result of dpl::copy");
 
     const std::vector<int> expectedCopy = {0, 2, 4, 6, 8, 10, 12, 14, 16, 18};
     EXPECT_EQ_N(expectedCopy.begin(), resultCopy.begin(), perm_size_result, "Wrong state of dpl::copy data");
 #endif // TEST_DPCPP_BACKEND_PRESENT
 
-    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+    return done(TEST_DPCPP_BACKEND_PRESENT);
 }
