+++ test/parallel_api/iterator/transform_iterator.pass.cpp
@@ -27,6 +27,8 @@ using namespace TestUtils;
 
 #if TEST_DPCPP_BACKEND_PRESENT
 
+#include "support/utils_sycl.h"
+
 DEFINE_TEST(test_copy)
 {
     DEFINE_TEST_CONSTRUCTOR(test_copy)
@@ -94,7 +96,7 @@ void test_simple_copy(size_t buffer_size)
     ::std::fill_n(host_source_begin, buffer_size, identity);
 
     test_copy<int> test(test_base_data);
-    TestUtils::invoke_on_all_hetero_policies<0>()(test, tr_sycl_source_begin, tr_sycl_source_end, sycl_result_begin, buffer_size, identity + 1);
+    invoke_on_all_hetero_policies<0>()(test, tr_sycl_source_begin, tr_sycl_source_end, sycl_result_begin, buffer_size, identity + 1);
 }
 
 void test_ignore_copy(size_t buffer_size)
@@ -126,7 +128,7 @@ void test_ignore_copy(size_t buffer_size)
     ::std::fill_n(host_result_begin, buffer_size, ignored);
 
     test_copy<int> test(test_base_data);
-    TestUtils::invoke_on_all_hetero_policies<1>()(test, sycl_source_begin, sycl_source_end, tr_sycl_result_begin, buffer_size, ignored);
+    invoke_on_all_hetero_policies<1>()(test, sycl_source_begin, sycl_source_end, tr_sycl_result_begin, buffer_size, ignored);
 }
 
 void test_multi_transform_copy(size_t buffer_size)
@@ -159,26 +161,33 @@ void test_multi_transform_copy(size_t buffer_size)
     ::std::fill_n(host_source_begin, buffer_size, identity);
 
     test_copy<int> test(test_base_data);
-    TestUtils::invoke_on_all_hetero_policies<2>()(test, tr3_sycl_source_begin, tr3_sycl_source_end, sycl_result_begin, buffer_size, identity + 3);
+    invoke_on_all_hetero_policies<2>()(test, tr3_sycl_source_begin, tr3_sycl_source_end, sycl_result_begin, buffer_size, identity + 3);
 }
 
 void
 test_copyable()
 {
-    //transform_iterator is not trivially_copyable, as it defines a copy assignment operator which does not copy
-    // its unary functor. This makes transform_iterator's device_copyable trait deprecated with sycl 2020.
-
-    EXPECT_TRUE(check_if_device_copyable_by_sycl2020_or_by_old_definition <oneapi::dpl::counting_iterator<int>>,
-                "counting_iterator is not device copyable");
+    static_assert(sycl::is_device_copyable_v<oneapi::dpl::counting_iterator<int>>,
+                  "counting_iterator is not device copyable");
 
     auto trans_count = ::dpl::make_transform_iterator(oneapi::dpl::counting_iterator<int>(0), [](auto i) { return i; });
-    EXPECT_TRUE(check_if_device_copyable_by_sycl2020_or_by_old_definition <decltype(trans_count)>,
-                "transform_iterator(counting_iterator) is not device copyable");
+    static_assert(sycl::is_device_copyable_v<decltype(trans_count)>,
+                  "transform_iterator is not device copyable with a counting iterator and noop lambda");
 
     std::vector<int> array(10, 0);
     auto trans_array = ::dpl::make_transform_iterator(array.begin(), [](auto i) { return i; });
-    EXPECT_TRUE(check_if_device_copyable_by_sycl2020_or_by_old_definition <decltype(trans_array)>,
-                "transform_iterator(host_iterator) is not device copyable");
+    static_assert(sycl::is_device_copyable_v<decltype(trans_array)>,
+                  "transform_iterator is not device copyable with a host iterator and noop lambda");
+
+    static_assert(sycl::is_device_copyable_v<
+                      oneapi::dpl::transform_iterator<constant_iterator_device_copyable, noop_device_copyable>>,
+                  "transform_iterator is not device copyable with device copyable types");
+
+    static_assert(!sycl::is_device_copyable_v<oneapi::dpl::transform_iterator<int*, noop_non_device_copyable>>,
+                  "transform_iterator is device copyable with non device copyable types");
+    static_assert(!sycl::is_device_copyable_v<
+                      oneapi::dpl::transform_iterator<constant_iterator_non_device_copyable, noop_device_copyable>>,
+                  "transform_iterator is device copyable with non device copyable types");
 }
 #endif // TEST_DPCPP_BACKEND_PRESENT
 
@@ -197,5 +206,5 @@ main()
     }
 #endif
 
-    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+    return done(TEST_DPCPP_BACKEND_PRESENT);
 }
